Pass Build on CircleCI and local Windows 10 VS2019
