+++ include/n2n.h
@@ -102,110 +102,12 @@
 
 struct ether_hdr
 {
-  uint8_t  dhost[ETH_ADDR_LEN];
-  uint8_t  shost[ETH_ADDR_LEN];
-  uint16_t type;                /* higher layer protocol encapsulated */
+    uint8_t  dhost[ETH_ADDR_LEN];
+    uint8_t  shost[ETH_ADDR_LEN];
+    uint16_t type;                /* higher layer protocol encapsulated */
 } __attribute__ ((__packed__));
 
 typedef struct ether_hdr ether_hdr_t;
-
-#if defined(__FreeBSD__) || defined(__NetBSD__) || defined(__OpenBSD__)
-#include <machine/endian.h>
-#endif
-
-#ifdef __OpenBSD__
-#include <endian.h>
-#define __BYTE_ORDER BYTE_ORDER
-#if BYTE_ORDER == LITTLE_ENDIAN
-#ifndef __LITTLE_ENDIAN__
-#define __LITTLE_ENDIAN__
-#endif /* __LITTLE_ENDIAN__ */
-#else
-#define __BIG_ENDIAN__
-#endif/* BYTE_ORDER */
-#endif/* __OPENBSD__ */
-
-
-#if __BYTE_ORDER == __LITTLE_ENDIAN
-#ifndef __LITTLE_ENDIAN__
-#define __LITTLE_ENDIAN__
-#endif
-#else
-#ifndef __BIG_ENDIAN__
-#define __BIG_ENDIAN__
-#endif
-#endif
-
-#ifdef WIN32
-#ifndef __LITTLE_ENDIAN__
-#define __LITTLE_ENDIAN__ 1
-#endif
-#endif
-
-#if !(defined(__LITTLE_ENDIAN__) || defined(__BIG_ENDIAN__))
-#if defined(__mips__)
-#undef __LITTLE_ENDIAN__
-#undef __LITTLE_ENDIAN
-#define __BIG_ENDIAN__
-#endif
-
-/* Everything else */
-#if (defined(__BYTE_ORDER__) && defined(__ORDER_LITTLE_ENDIAN__))
-#if __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__
-#define __LITTLE_ENDIAN__
-#else
-#define __BIG_ENDIAN__
-#endif
-#endif
-
-#endif
-
-/* *************************************** */
-
-struct n2n_iphdr {
-#if defined(__LITTLE_ENDIAN__)
-  u_int8_t ihl:4, version:4;
-#elif defined(__BIG_ENDIAN__)
-  u_int8_t version:4, ihl:4;
-#else
-# error "Byte order must be defined"
-#endif
-  u_int8_t tos;
-  u_int16_t tot_len;
-  u_int16_t id;
-  u_int16_t frag_off;
-  u_int8_t ttl;
-  u_int8_t protocol;
-  u_int16_t check;
-  u_int32_t saddr;
-  u_int32_t daddr;
-} __attribute__ ((__packed__));
-
-struct n2n_tcphdr
-{
-  u_int16_t source;
-  u_int16_t dest;
-  u_int32_t seq;
-  u_int32_t ack_seq;
-#if defined(__LITTLE_ENDIAN__)
-  u_int16_t res1:4, doff:4, fin:1, syn:1, rst:1, psh:1, ack:1, urg:1, ece:1, cwr:1;
-#elif defined(__BIG_ENDIAN__)
-  u_int16_t doff:4, res1:4, cwr:1, ece:1, urg:1, ack:1, psh:1, rst:1, syn:1, fin:1;
-#else
-# error "Byte order must be defined"
-#endif
-  u_int16_t window;
-  u_int16_t check;
-  u_int16_t urg_ptr;
-} __attribute__ ((__packed__));
-
-struct n2n_udphdr
-{
-  u_int16_t source;
-  u_int16_t dest;
-  u_int16_t len;
-  u_int16_t check;
-} __attribute__ ((__packed__));
 
 #ifdef HAVE_LIBZSTD
 #include <zstd.h>
@@ -400,7 +302,6 @@ typedef struct filter_rule
 // for impl, see: network_traffic_filter.c
 uint8_t process_traffic_filter_rule_str(const char* rule_str, filter_rule_t* rule_struct);
 
-#ifdef FILTER_TRAFFIC
 /*
  * network traffic filter interface
  */
@@ -418,7 +319,6 @@ typedef struct network_traffic_filter
     n2n_verdict (*filter_packet_from_tap)(struct network_traffic_filter* filter, n2n_edge_t *eee, uint8_t *payload, uint16_t payload_size);
 
 } network_traffic_filter_t;
-#endif
 
 /* *************************************************** */
 
@@ -462,9 +362,7 @@ typedef struct n2n_edge_conf {
   int                 register_ttl;           /**< TTL for registration packet when UDP NAT hole punching through supernode. */
   int                 local_port;
   int                 mgmt_port;
-#ifdef FILTER_TRAFFIC
   filter_rule_t       *network_traffic_filter_rules;
-#endif
 } n2n_edge_conf_t;
 
 
@@ -518,9 +416,7 @@ struct n2n_edge {
 
   n2n_tuntap_priv_config_t tuntap_priv_conf;   /**< Tuntap config */
 
-#ifdef FILTER_TRAFFIC
   network_traffic_filter_t *network_traffic_filter;
-#endif
 };
