+++ src/edge_utils.c
@@ -271,10 +271,8 @@ n2n_edge_t* edge_init(const n2n_edge_conf_t *conf, int *rv) {
     goto edge_init_error;
   }
 
-#ifdef FILTER_TRAFFIC
   eee->network_traffic_filter = create_network_traffic_filter();
   network_traffic_filter_add_rule(eee->network_traffic_filter, eee->conf.network_traffic_filter_rules);
-#endif
 
   //edge_init_success:
   *rv = 0;
@@ -1180,13 +1178,11 @@ static int handle_PACKET(n2n_edge_t * eee,
 	}
       }
 
-#ifdef FILTER_TRAFFIC
       if(eee->network_traffic_filter->filter_packet_from_peer( eee->network_traffic_filter, eee, orig_sender,
 								eth_payload, eth_size ) == N2N_DROP){
         traceEvent(TRACE_DEBUG, "Filtered packet %u", (unsigned int)eth_size);
         return(0);
       }
-#endif
 
       if(eee->cb.packet_from_peer) {
 	uint16_t tmp_eth_size = eth_size;
@@ -1738,7 +1734,6 @@ void edge_read_from_tap(n2n_edge_t * eee) {
         }
       else
         {
-#ifdef FILTER_TRAFFIC
 	  if(eee->network_traffic_filter) {
 	    if( eee->network_traffic_filter->filter_packet_from_tap( eee->network_traffic_filter, eee, eth_pkt,
 								     len) == N2N_DROP){
@@ -1746,7 +1741,6 @@ void edge_read_from_tap(n2n_edge_t * eee) {
               return;
 	    }
 	  }
-#endif
 
 	  if(eee->cb.packet_from_tap) {
 	    uint16_t tmp_len = len;
@@ -2045,8 +2039,8 @@ void readFromIPSocket(n2n_edge_t * eee, int in_sock) {
                   }
 
                   /* REVISIT: find a more elegant expression to increase following pointers. */
-                  tmp_sock = (void*)tmp_sock + REG_SUPER_ACK_PAYLOAD_ENTRY_SIZE;
-                  tmp_mac = (void*)tmp_sock + sizeof(n2n_sock_t);
+                  tmp_sock = (char*)tmp_sock + REG_SUPER_ACK_PAYLOAD_ENTRY_SIZE;
+                  tmp_mac = (char*)tmp_sock + sizeof(n2n_sock_t);
 		}
 
 		eee->last_sup = now;
@@ -2319,9 +2313,7 @@ void edge_term(n2n_edge_t * eee) {
 
   edge_cleanup_routes(eee);
 
-#ifdef FILTER_TRAFFIC
   destroy_network_traffic_filter(eee->network_traffic_filter);
-#endif
 
   closeTraceFile();
 
@@ -2822,7 +2814,6 @@ void edge_term_conf(n2n_edge_conf_t *conf) {
   if (conf->routes) free(conf->routes);
   if (conf->encrypt_key) free(conf->encrypt_key);
 
-#ifdef FILTER_TRAFFIC
   if(conf->network_traffic_filter_rules)
     {
       filter_rule_t *el = 0, *tmp = 0;
@@ -2832,7 +2823,6 @@ void edge_term_conf(n2n_edge_conf_t *conf) {
 	  free(el);
 	}
     }
-#endif
 }
 
 /* ************************************** */
